requirements.txt
asgiref==3.2.3
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.2.1
django-extensions==2.2.8
django-heroku==0.3.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
idna==2.8
Pillow==7.0.0
psycopg2==2.8.4
pycparser==2.19
pyfcm==1.4.7
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
gunicorn